+++ src/main/java/com/demoJob/demo/entity/Job.java
@@ -30,7 +30,7 @@ public class Job extends AbstractEntity<Long> {
     private String location;
 
     @Min(value = 0, message = "Lương không được âm")
-    private BigDecimal salary;
+    private double salary;
 
     @Min(value = 1, message = "Số lượng phải lớn hơn 0")
     private int quantity;
